Make all pages a single file components
