--- src/pages/AddApp/AddApp.scss
@@ -1,51 +1,0 @@
-@import "~@aeternity/aepp-components/dist/variables.scss";
-
-.add-app.mobile-page {
-  /deep/ .panel {
-    margin-top: 0;
-  }
-
-  .ae-button._size_smaller {
-    padding-right: 0;
-    padding-left: 0;
-
-    .ae-icon {
-      margin-left: 0;
-      margin-right: 6px;
-    }
-  }
-
-  .add-button-wrapper {
-    text-align: right;
-    margin-top: -20px;
-    margin-bottom: 20px;
-  }
-
-  .app {
-    display: flex;
-    cursor: pointer;
-
-    &.inactive {
-      cursor: not-allowed;
-    }
-
-    .content {
-      display: flex;
-      flex-direction: column;
-      justify-content: space-around;
-      margin-left: 15px;
-
-      &:before, &:after {
-        content: '';
-      }
-
-      h2 {
-        margin: 0;
-      }
-    }
-  }
-
-  .ae-divider {
-    margin: 16px 0;
-  }
-}
