--- src/pages/AppBrowser/AppBrowser.scss
@@ -1,53 +1,0 @@
-.app-browser {
-  overflow: hidden;
-
-  iframe, .loader {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-  }
-
-  iframe {
-    border: none;
-    transition: filter 2s;
-
-    &.loading {
-      filter: blur(3px);
-    }
-  }
-
-  .loader {
-    background-color: rgba(0,0,0,0.15);
-    cursor: wait;
-
-    @keyframes sk-bounce {
-      0%, 100% {
-        transform: scale(0.0);
-      }
-      50% {
-        transform: scale(1.0);
-      }
-    }
-
-    &:before, &:after {
-      content : '';
-      display: block;
-      position: absolute;
-      top: 30%;
-      $circleSide: 60px;
-      left: calc(50% - #{$circleSide / 2});
-      height: $circleSide;
-      width: $circleSide;
-      background-color: #F03C6E;
-      opacity: 0.6;
-      border-radius: 50%;
-      animation: sk-bounce 2.0s infinite ease-in-out;
-    }
-
-    &:after {
-      animation-delay: -1s;
-    }
-  }
-}
