--- src/pages/Apps/Apps.scss
@@ -1,112 +1,0 @@
-@import "~@aeternity/aepp-components/dist/mixins.scss";
-
-.apps {
-  @include abovePhone {
-    display: flex;
-    flex-direction: column;
-    position: fixed;
-    top: 0;
-    right: 0;
-    bottom: 0;
-    left: 0;
-    padding: 30px 0 80px 0;
-    overflow-y: auto;
-  }
-
-  .header-mobile {
-    @include abovePhone {
-      display: none;
-    }
-  }
-
-  .header-desktop {
-    position: fixed;
-    top: 0;
-    right: 0;
-    left: 0;
-
-    @include phone {
-      display: none;
-    }
-
-    .ae-button .ae-icon {
-      width: 24px;
-      height: 24px;
-      margin: 13px;
-    }
-  }
-
-  .shortcuts {
-    display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(90px, 1fr));
-    margin: 50px 10px 10px 10px;
-    grid-gap: 30px 10px;
-    justify-items: center;
-
-    @include abovePhone {
-      grid-gap: #{100 / 1024 * 100vh} #{114 / 1440 * 100vw};
-      margin: auto #{270 / 1440 * 100vw};
-    }
-
-    .app-shortcut {
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-      width: 100%;
-
-      .app-icon-wrapper {
-        position: relative;
-
-        @mixin remove-app-btn-hidden {
-          visibility: hidden;
-          opacity: 0;
-          transform: translate(48%, -50%) scale(.45, .45);
-          transition: opacity 300ms, transform 300ms, visibility 0s 300ms;
-        }
-
-        .remove-app-btn {
-          position: absolute;
-          top: 0;
-          right: 0;
-          cursor: pointer;
-          @include remove-app-btn-hidden;
-        }
-
-        &:hover .remove-app-btn, .remove-app-btn.visible {
-          visibility: visible;
-          opacity: 1;
-          transform: translate(48%, -50%) scale(.58, .58);
-          transition: opacity 300ms 500ms, transform 300ms 500ms, visibility 0s 500ms;
-        }
-
-        @media (hover: none) {
-          .remove-app-btn:not(.visible) {
-            @include remove-app-btn-hidden;
-          }
-        }
-
-        .ae-app-icon {
-          margin: 0 auto;
-        }
-      }
-
-      .app-name {
-        display: block;
-        margin-top: 13px;
-        font-size: 18px;
-        font-weight: 500;
-        text-align: center;
-        width: 100%;
-        overflow: hidden;
-        white-space: nowrap;
-        text-overflow: ellipsis;
-      }
-
-      @include abovePhone {
-        &.add-app {
-          display: none;
-        }
-      }
-    }
-  }
-}
