--- src/pages/Transfer/Transfer.scss
@@ -1,59 +1,0 @@
-@import '~@aeternity/aepp-components/dist/variables.scss';
-
-.transfer.mobile-page {
-  .swiper-container /deep/ {
-    z-index: auto;
-
-    .swiper-wrapper {
-      z-index: auto;
-      padding-bottom: 35px;
-    }
-
-    .swiper-pagination {
-      z-index: auto;
-
-      &-bullet {
-        background: $aubergine;
-      }
-    }
-  }
-
-  .ae-switch {
-    margin-bottom: 35px;
-    text-transform: uppercase;
-  }
-
-  .buttons {
-    display: flex;
-    justify-content: space-between;
-    margin-top: -15px;
-    margin-bottom: 15px;
-
-    .ae-button {
-      width: auto;
-      min-width: 25%;
-      margin: 0;
-    }
-  }
-
-  .fiat-amount {
-    color: $grey;
-    text-align: center;
-    font-size: 12px;
-    line-height: 16px;
-    font-weight: 500;
-    margin-top: -20px;
-    margin-bottom: 20px;
-    text-transform: uppercase;
-  }
-
-  .send-button {
-    margin-bottom: 80px;
-
-    img {
-      height: 20px;
-      margin-left: 50px;
-      margin-right: -50px;
-    }
-  }
-}
